lesson3/requirements.txt
GitPython==2.1.7
nltk==3.2.4

